requirements.txt
aiogram==3.14.0
brawlstats==4.2.0
Pillow==11.0.0
python-dotenv==1.0.1
Requests==2.32.3
